Refactoring of Functions
