+++ core/src/main/java/io/doov/core/dsl/field/BaseFieldInfo.java
@@ -24,7 +24,7 @@ import java.util.function.*;
 
 import io.doov.core.dsl.DslField;
 import io.doov.core.dsl.impl.DefaultCondition;
-import io.doov.core.dsl.impl.StringFunction;
+import io.doov.core.dsl.impl.base.StringFunction;
 import io.doov.core.dsl.impl.num.IntegerFunction;
 import io.doov.core.dsl.lang.StepCondition;
