+++ core/src/main/java/io/doov/core/dsl/field/types/IterableFieldInfo.java
@@ -19,7 +19,7 @@ import io.doov.core.FieldInfo;
 import io.doov.core.dsl.field.BaseFieldInfo;
 import io.doov.core.dsl.field.DelegatingFieldInfoImpl;
 import io.doov.core.dsl.impl.DefaultCondition;
-import io.doov.core.dsl.impl.IterableFunction;
+import io.doov.core.dsl.impl.base.IterableFunction;
 import io.doov.core.dsl.lang.StepCondition;
 
 public class IterableFieldInfo<T, C extends Iterable<T>> extends DelegatingFieldInfoImpl implements BaseFieldInfo<C> {
